Update with js from file
