+++ src/routes/+page.svelte
@@ -1,36 +1,19 @@
 <script>
     import {onMount} from 'svelte';
-
-    let name = "alBz";
-    let role = "Senior Fullstack Developer";
-    let cv_filename = "albertobarrago_cv_2024.pdf"
-    let githubRepoString = "https://github.com/AlbertoBarrago?tab=repositories"
-    let tagline = "Building digital experiences that matter"
-    let skills = [
-        "Frontend Development",
-        "Backend Architecture",
-        "Cloud Solutions",
-        "Team Leadership"
-    ];
-    let version = "v1.2.0"
-    let email = "albertobarrago@gmail.com";
-
-
-    let isVisible = false;
-
-    const downloadCv = () => {
-        const link = document.createElement('a');
-        link.href = cv_filename;
-        link.download = cv_filename;
-        link.dispatchEvent(new MouseEvent('click'));
-    }
-
-    const mailtoLink = `mailto:${email}`;
+    import {
+        isVisible,
+        version,
+        name,
+        skills,
+        tagline,
+        role,
+        githubRepoString,
+        mailtoLink,
+        downloadCv,
+    } from '../js/index.js'
 
     onMount(() => {
-        isVisible = true;
         particlesJS.load('particles-js', './particles-config.json');
-
         const script = document.createElement('script');
         script.src = '/oneko.js';
         script.async = true;
@@ -69,7 +52,6 @@
             <small>{version}</small>
         </footer>
     </section>
-
 </main>
